August 7th Update
Esc button now opens a pause menu in the game scene.

pausing the game does;
unlocks cursor
freezes game with timescale = 0

Also locked the cursor when game starts so the cursor does not fly out of the window while moving the camera. this can be controlled with the L key or Esc for the pause menu